requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
asgiref==3.10.0
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
billiard==4.2.2
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
django==5.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
dotenv==0.9.9
    # via -r requirements.in
gigachain==100.0.4
    # via -r requirements.in
gigachat==0.1.42.post2
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   gigachat
    #   python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
kombu==5.5.4
    # via celery
packaging==25.0
    # via
    #   gunicorn
    #   kombu
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.3
    # via gigachat
pydantic-core==2.41.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   dotenv
python-telegram-bot==22.5
    # via -r requirements.in
redis==7.0.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.15.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via
    #   django
    #   kombu
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
